+++ packages/homepage/.eslintrc
@@ -2,5 +2,8 @@
   "globals": {
     "graphql": true
   },
-  "rules": { "jsx-a11y/href-no-hash": 0 }
+  "rules": {
+    "jsx-a11y/href-no-hash": 0,
+    "react/no-unescaped-entities": 0
+  }
 }
