+++ packages/homepage/src/screens/home/Users/index.js
@@ -3,15 +3,7 @@ import styled from 'styled-components';
 import { StaticQuery, graphql, Link } from 'gatsby';
 
 import MaxWidth from 'common/components/flex/MaxWidth';
-
-function shuffleArray(array) {
-  for (let i = array.length - 1; i > 0; i--) {
-    const j = Math.floor(Math.random() * (i + 1));
-    [array[i], array[j]] = [array[j], array[i]];
-  }
-
-  return array;
-}
+import shuffleArray from '../../../utils/shuffleArray';
 
 const Container = styled.div`
   background-color: ${({ theme }) => theme.background2};
